+++ source/Processing.IntegrationTests/Application/Customers/GetCustomerMasterData/SampleData.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Processing.IntegrationTests.Factories;
+
 namespace Processing.IntegrationTests.Application.Customers.GetCustomerMasterData
 {
     public static class SampleData
