+++ source/Processing.IntegrationTests/Application/MoveIn/ProcessPendingProcessesTests.cs
@@ -17,6 +17,7 @@ using MediatR;
 using Processing.Application.Common.TimeEvents;
 using Processing.Domain.MeteringPoints;
 using Processing.Domain.SeedWork;
+using Processing.IntegrationTests.Factories;
 using Processing.IntegrationTests.Fixtures;
 using Processing.IntegrationTests.TestDoubles;
 using Xunit;
